+++ R/friskvik.R
@@ -108,4 +108,47 @@ do_filter_friskvik_tabs <- function(dt, dscr){
   return(dt)
 }
 
+#' @title generate_specific_friskvik_indicators
+#' @param cubename name of cube
+#' @param friskvik_id optional, specify the ID of the indicators you want to generate
+#' @param year year, defaults to getOption("khfunctions.year")
+#' @export
+#' @examples
+#' # implement_specific_friskvik_indicators(cubename = "UTDN", friskvik_id = c(1719, 1771), year = 2025) # Specific indicators and year
+#' # implement_specific_friskvik_indicators(cubename = "UTDN") # All indicators, default production year
+generate_specific_friskvik_indicators <- function(cubename = NULL, friskvik_id = NULL, year = getOption("khfunctions.year")){
+  on.exit(RODBC::odbcCloseAll())
+  if(is.null(cubename)) stop("cubename must be provided, cannot be NULL")
+  overwritewarning <- "\n** Files are overwritten if they already exist.\n\n*** NB! Only csv folder is affected, not preexisting godkjent folders!!"
 
+  user_args <- as.list(environment())
+  user_args[["name"]] <- cubename
+  capture.output(parameters <- get_cubeparameters(user_args = user_args))
+  valid_cube <- read_kubestatus(parameters$dbh, year)[KUBE_NAVN == cubename]
+  if(nrow(valid_cube) == 0){
+    cat("\n*Ingen godkjent kube funnet i KUBESTATUS, kan ikke lage friskvikfiler")
+    return(invisible(NULL))
+  }
+ 
+  indicators <- parameters$friskvik[, .SD, .SDcols = c("INDIKATOR", "ID")]
+  if(is.null(friskvik_id)){
+    cat("\n* Generating all friskvik files as 'friskvik_id = NULL', ID(s):", paste(indicators$ID, collapse = ", "), overwritewarning)
+  } else {
+    if(any(!friskvik_id %in% indicators$ID)) stop("At least 1 of the requested friskvik_id(s) (", paste(friskvik_id, collapse = ", "), ") does not exist!")
+    indicators <- indicators[ID %in% friskvik_id]
+    cat("\n* Generating requested friskvik ID(s):", paste(indicators$ID, collapse = ", "), overwritewarning)
+  }
+  
+  continue <- utils::menu(c("Yes, continue!", "No, stop!"), title = "\n\nIs it ok to generate (AND REPLACE) the requested friskvik indicator(s)?")
+  if(continue == 2) return(invisible(NULL))
+  
+  cube_name <- paste0(valid_cube$KUBE_NAVN, "_", valid_cube$DATOTAG_KUBE, ".rds")
+  path <- file.path(getOption("khfunctions.root"), getOption("khfunctions.kubedir"), getOption("khfunctions.kube.dat"), "R", cube_name)
+  if(!file.exists(path)) stop("Can't find cube file ", path, "\n Check if date tag in KUBESTATUS is correct!!")
+  cube <- data.table::setDT(readRDS(path))
+  do_remove_censored_observations(dt = cube, outvalues = get_outvalues_allvis(parameters = parameters))
+  
+  for(i in 1:nrow(indicators)){
+    generate_friskvik_indicator(dt = cube, id = indicators[i, ID], parameters = parameters)
+  }
+}