+++ R/special_handling.R
@@ -18,11 +18,14 @@ do_special_handling <- function(name, dt, code, parameters, koblid = NULL){
   is_stata <- grepl("<STATA>", code)
   
   if(is_stata){
+    cat("\n** Starter STATA-snutt:", name)
     code <- gsub("<STATA>[ \n]*(.*)", "\\1", code)
     dt <- do_stata_processing(dt = dt, script = code, parameters = parameters)
+    cat("\n** Ferdig i STATA")
     return(dt)
   }
   
+  cat("\n** Starter R-snutt:", name)
   code_env <- new.env()
   dtname <- as.character(substitute(dt))
   assign(dtname, dt, envir = code_env)
@@ -34,6 +37,7 @@ do_special_handling <- function(name, dt, code, parameters, koblid = NULL){
   }
   assign(dtname, get(dtname, envir = code_env), envir = code_env)
   dt <- get(dtname, envir = code_env)
+  cat("\n** R-snutt ferdig")
   return(dt)
 }
 
@@ -50,26 +54,43 @@ do_special_handling <- function(name, dt, code, parameters, koblid = NULL){
 #' @param batchdate used to generate file names
 #' @param stata_exe path to STATA program
 do_stata_processing <- function(dt, script, parameters){
+  on.exit(stata_processing_cleanup(statafiles = statafiles, orgwd = orgwd), add = T)
   tmpdir <- file.path(fs::path_home(), "helseprofil", "STATAtmp")
   if(!fs::dir_exists(tmpdir)) fs::dir_create(tmpdir)
-  batchdate <- parameters$batchdate
   orgwd <- getwd()
   setwd(tmpdir)
+  batchdate <- parameters$batchdate
+  use_parquet <- can_use_parquet(dt = dt)
   statafiles <- set_stata_filenames(batchdate = batchdate, tmpdir = tmpdir)
   charactercols <- names(dt)[sapply(dt, is.character)]
   dt[, (charactercols) := lapply(.SD, function(x) ifelse(x == "", " ", x)), .SDcols = charactercols] 
-  fix_column_names_pre_stata(dt = dt)
-  haven::write_dta(dt, statafiles$dta)
-  on.exit(file.remove(statafiles$dta), add = T)
-  generate_stata_do_file(script = script, statafiles = statafiles)
+  statanames <- fix_column_names_pre_stata(oldnames = names(dt))
+  data.table::setnames(dt, statanames)
+  
+  do_write_stata_file(dt = dt, statafiles = statafiles, use_parquet = use_parquet)
+  generate_stata_do_file(script = script, statafiles = statafiles, use_parquet = use_parquet)
   run_stata_script(dofile = statafiles$do, stata_exe = parameters$StataExe)
   check_stata_log_for_error(statafiles = statafiles)
-  dt <- data.table::setDT(haven::read_dta(statafiles$dta))
+  dt <- do_read_stata_file(statafiles = statafiles, use_parquet = use_parquet)
+  
+  charactercols <- names(dt)[sapply(dt, is.character)]
   dt[, (charactercols) := lapply(.SD, function(x) ifelse(x == " ", "", x)), .SDcols = charactercols] 
-  fix_column_names_post_stata(dt = dt)
-  # if(!is(dt, "data.table")) data.table::setDT(dt)
-  setwd(orgwd)
+  rnames <- fix_column_names_post_stata(oldnames = names(dt))
+  data.table::setnames(dt, rnames)
   return(dt)
+}
+
+#' @descriptio Delete temporary data files and reset working directory
+#' @keywords internal
+#' @noRd
+stata_processing_cleanup <- function(statafiles, orgwd){
+  cat("\n*** Sletter midlertidige datafiler")
+  gc()
+  for(file in c("parquet_in", "parquet_out", "dta")){
+    path <- statafiles[[file]]
+    if(file.exists(path)) file.remove(path)
+  }
+  setwd(orgwd)
 }
 
 #' @description 
@@ -77,24 +98,49 @@ do_stata_processing <- function(dt, script, parameters){
 #' Some specific conversions are later reversed when data is read back into R.
 #' Some general conversions are not reversed. 
 #' @noRd
-fix_column_names_pre_stata <- function(dt){
-  fixednames <- names(dt)
-  fixednames <- gsub("^(\\d.*)$", "S_\\1", fixednames)
+fix_column_names_pre_stata <- function(oldnames){
+  cat("\n*** Fikser kolonnenavn før stata")
+  fixednames <- oldnames
+  fixednames <- gsub("^(\\d.*)$", "S_\\1", fixednames, perl = TRUE)
   fixednames <- gsub("^(.*)\\.(f|a|n|fn1|fn3|fn9)$", "\\1_\\2", fixednames)
-  
-  fixednames <- gsub("[^a-zA-Z0-9_æÆøØåÅ]", "_", fixednames)
+  fixednames <- gsub("[^[:alnum:]_\u00c6\u00d8\u00c5\u00e6\u00f8\u00e5]", "_", fixednames, perl = TRUE)
   fixednames <- gsub("^(?![a-zA-Z_])(.*)", "_\\1", fixednames, perl = TRUE)
   fixednames <- gsub("^_+$", "var", fixednames)
   fixednames <- substr(fixednames, 1, 32)
-  data.table::setnames(dt, old = names(dt), new = fixednames)
+  return(fixednames)
 }
 
 #' @noRd
-fix_column_names_post_stata <- function(dt){
-  fixednames <- names(dt)
+fix_column_names_post_stata <- function(oldnames){
+  cat("\n*** Leser filen inn igjen og fikser kolonnenavn")
+  fixednames <- oldnames
   fixednames <- gsub("^S_(\\d.*)$", "\\1", fixednames)
   fixednames <- gsub("^(.*)_(f|a|n|fn1|fn3|fn9)$", "\\1.\\2", fixednames)
-  data.table::setnames(dt, fixednames)
+  return(fixednames)
+}
+
+#' @title can_use_parquet
+#' @description
+#' If the longest value of a character column contain [æøå], represented by unicode sequences.
+#' STATA will crop the value when reading a .parquet file, as these characters are multibyte. 
+#' This function returns TRUE if this is not the case, indicating that .parquet can be used.
+#' If the longest value of any column contain [æøå], FALSE is returned, indicating that .dta-format should be used. 
+#' @param dt data
+#' @keywords internal
+#' @noRd
+can_use_parquet <- function(dt){
+  out <- TRUE
+  charcols <- names(dt)[sapply(dt, is.character)]
+  i <- 1
+  while(out && i <= length(charcols)){
+    vals <- collapse::funique(dt[[charcols[i]]])
+    vals <- vals[nchar(vals) == collapse::fmax(nchar(vals))]
+    if(any(grepl("[\u00c6\u00d8\u00c5\u00e6\u00f8\u00e5]", vals))){
+      out <- FALSE
+    }
+    i <- i + 1
+  }
+  return(out)
 }
 
 #' @noRd
@@ -103,20 +149,66 @@ set_stata_filenames <- function(batchdate, tmpdir){
   statafiles[["do"]] <-  file.path(tmpdir, paste0("STATAtmp_", batchdate, ".do"))
   statafiles[["dta"]] <- file.path(tmpdir, paste0("STATAtmp_", batchdate, ".dta"))
   statafiles[["log"]] <- file.path(tmpdir, paste0("STATAtmp_", batchdate, ".log"))
+  statafiles[["parquet_out"]] <- file.path(tmpdir, paste0("STATAtmp_out", batchdate, ".parquet"))
+  statafiles[["parquet_in"]] <- file.path(tmpdir, paste0("STATAtmp_in", batchdate, ".parquet"))
   return(statafiles)
 }
 
+#' @title do_write_stata_file
+#' @description
+#' Write file to be processed by stata.
+#' If use_parquet, the statafiles$parquet_out file is written. If not, the statafiles$dta file is used.
+#' @param dt data
+#' @param statafiles list of paths
+#' @param use_parquet TRUE/FALSE
+#' @keywords internal
 #' @noRd
-generate_stata_do_file <- function(script, statafiles){
-  sink(statafiles$do)
-  cat("use ", statafiles$dta, "\n", sep = "")
-  cat(script, "\n")
-  cat("save ", statafiles$dta, ", replace\n", sep = "")
-  sink()
+do_write_stata_file <- function(dt, statafiles, use_parquet){
+  cat("\n*** Skriver STATA-fil")
+  if(use_parquet){
+    do_write_parquet(dt = dt, filepath = statafiles$parquet_out)
+  } else {
+    haven::write_dta(dt, statafiles$dta)
+  }
+}
+
+#' @title do_read_stata_file
+#' @description
+#' Reads in file after stata processing. If use_parquet, the statafiles$parquet_in file is read, or the statafiles$dta file is read.
+#' @param statafiles list of paths
+#' @param use_parquet TRUE/FALSE
+#' @keywords internal
+#' @noRd
+do_read_stata_file <- function(statafiles, use_parquet){
+  if(use_parquet){
+    dt <- data.table::copy(data.table::as.data.table(arrow::read_parquet(statafiles$parquet_in)))
+  } else {
+    dt <- data.table::setDT(haven::read_dta(statafiles$dta))
+  }
+  return(dt)
+}
+
+#' @noRd
+generate_stata_do_file <- function(script, statafiles, use_parquet){
+  if(use_parquet){
+    sink(statafiles$do)
+    cat("net install pq, from(http://fmwww.bc.edu/RePEc/bocode/p)\n") 
+    cat("pq use using ", statafiles$parquet_out, "\n", sep = "")
+    cat(script, "\n")
+    cat("pq save using ", statafiles$parquet_in, ", replace\n", sep = "")
+    sink()
+  } else {
+    sink(statafiles$do)
+    cat("use ", statafiles$dta, "\n", sep = "")
+    cat(script, "\n")
+    cat("save ", statafiles$dta, ", replace\n", sep = "")
+    sink()
+  }
 }
 
 #' @noRd
 run_stata_script <- function(dofile, stata_exe){
+  cat("\n*** Kjører STATA-script...")
   call <- paste("\"", stata_exe, "\" /e do ", dofile, " \n", sep = "")
   system(call, intern = TRUE)
 }
@@ -131,5 +223,4 @@ check_stata_log_for_error <- function(statafiles){
   }
   return(invisible(NULL))
 }
-